Fix race condition when syncing shift details
Previously, we fetched summaries using `select_for_update` and related
details using `prefetch_related` with the intent of serializing
concurrent jobs and preventing duplicate requests.

However, because `select_related` issues a join query, the summaries and
details are fetched in a single query. This query blocked correctly
because of the `select_for_update`, but after unblocking did not see the
details added by the previous job.

Changing `select_related` to `prefetch_related` issues a separate query
instead, and this is able to pick up the newly-synced details and avoid
trying to sync them again.